dior: Move hardware specific .rc files to /vendor/etc/init/hw
The reason these files are moved to /vendor/etc/init/hw and
not /vendor/etc/init is because 'init' scans the /vendor/etc/init folder
for vendor service .rc fragments to import. That would either result
in duplicate imports or out-of-order imports of these hardware
specific .rc files and that *may* have unintended consequences.

In order to preserve the current import order for MTP, we move all
hardware specific .rc files to /vendor/etc/init/hw which is not
autoscanned by init BUT leave the init.${ro.hardware}.rc (i.e.
init.qcom.rc in this case) in root for 'init' to find.

https://github.com/haran2003/android_device_xiaomi_armani/commit/997d6aaa5dee5da26112b2c4fc7e4283c1c6effc

Change-Id: If712df704026340fb3ffc26492c5bd00276b4ddf